+++ src/components/gameslist/components/AutomateButtons.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Button } from '@nextui-org/react';
+import { TbCardsFilled } from "react-icons/tb";
 import { FaAward } from 'react-icons/fa';
-import { IoPlay } from 'react-icons/io5';
 import { useAutomate } from '../hooks/useAutomateButtons';
 
 export default function AutomateButtons() {
@@ -13,7 +13,7 @@ export default function AutomateButtons() {
                 size='sm'
                 color='primary'
                 className='rounded-full font-semibold'
-                startContent={<IoPlay fontSize={15} />}
+                startContent={<TbCardsFilled fontSize={18} />}
                 onPress={startCardFarming}
             >
                 Card farming