+++ changelog.md
@@ -1,6 +1,5 @@
 [![Downloads](https://img.shields.io/github/downloads/zevnda/steam-game-idler/1.6.2/total?style=for-the-badge&logo=github&color=137eb5)](https://github.com/zevnda/steam-game-idler/releases/download/1.6.2/Steam.Game.Idler_1.6.2_x64_en-US.msi)
 
 ## Changelog
-- New logo/icon for both the SGI and SteamUtility windows
-  - If anyone is good at making custom logos/icons, let me know
-- Fixed the height of the draggable area of the `titlebar` when on the `settings` screen
+- Added a guided tutorial of the UI that will be displayed when launching SGI for the first time
+- Fixed an issue causing the timer on the `card farming` screen to be set multiple times when farming more than one game